requirements.txt
requests
python-telegram-bot
aiohttp
yfinance
feedparser